Initial work & notes on contextual bindings
I no longer believe overwriting w/W/b/B/e is a good idea, despite the
similiarity in what a "motion" means within scheme files.

How should `b` be handled when inside a string, or a comment? What about
if `b` would land on a string, or comment? Should it go to the start of
the string s-expression, or the beginning of the final vim-"word" of
that string?

Really just committing my initial thinkies and work as I consider
context-based motions.