Finish reworking motions to only use binary search
I'm beginning to believe that contextual bindings simply are not
possible. Too many sharp edges to work around. Should doc comments be
handled the same as regular comments. Should strings be handled as
comments, or as code elements. Etc.

Easier to have s-expression motions be standalone, and regular vim
motions be constant.

After all, editing should be predictable, and expected. Should not need
to maintain a contextual understanding of where you are in the text,
where a "previous" jump would place the cursor, and handle the motion
depending on a potential jump target.


cursor.lua
   Fix bug w/ `is_behind`, was also including the first column of the
   node itself.


predicates.lua
   Add `is_string`, though now I think I'm going to remove this as I
   take out the rest of the contextual stuff.